Install rpm packages from yum-puppetcore.puppet.com
Change the default yum_source to https://yum-puppetcore.puppet.com. Note the
change to HTTPS.

Add optional username and password to install task. If installing from the
default yum_source, then the username defaults to forge-key and a password
must be specified.

For SLES, add the credentials to /etc/zypp/credentials.d/PuppetcoreCreds

For other rpm platforms, add the credentials to the baseurl.

Create dnf and sles Dockerfiles for testing the install task.

Create install.sh script to build docker image and run it:

    docker/bin/install.sh [image] [version]

By default install 8.11.0 on rocky8.

The `PUPPET_FORGE_TOKEN` environment variable must be set, which will be passed
as the `password` to the task.