Upgrade rpm packages from yum-puppetcore.puppet.com
Add optional username and password parameters to the `puppet_agent` class.

If `manage_repo` is true, then add the credentials to the repo config (for RPM
platforms other than SLES) with secure permissions.

For SLES, add credentials to /etc/zypp/credentials.d/PuppetcoreCreds with secure
permissions. Also include auth=basic and credentials=PuppetcoreCreds to the
baseurl.

Update the Dockerfile to install 7.34.0 from yum.puppet.com and upgrade to
8.11.0 from yum-puppetcore, to verify the module can upgrade agents on amazon
2023, fedora 40, rocky 8 and sles 15.

    export PUPPET_FORGE_TOKEN=...
    docker/bin/upgrade.sh [platform] [from] [to]

where platform is one of amazon, fedora, rocky or sles and from/to are
puppet-agent versions.

The password is passed to the `docker run` command as an environment variable,
so that it's not persisted in the docker image.